requirements.txt
# Azure Cognitive Services
azure-ai-formrecognizer>=3.3.0
azure-cognitiveservices-vision-computervision>=0.9.0
msrest>=0.7.1

# OpenAI pour les embeddings
openai>=1.0.0

# Supabase
supabase>=2.0.0

# Traitement de données
pandas>=2.0.0
numpy>=1.24.0

# Images
Pillow>=10.0.0
pdf2image>=1.16.0

# Utilitaires
python-dotenv>=1.0.0
tqdm>=4.65.0
tenacity>=8.2.0

# Logging
colorlog>=6.7.0
